refactor: move startup logic files to startup dir
