+++ packages/neuron-wallet/src/main.ts
@@ -1,8 +1,8 @@
 import { app, Menu } from 'electron'
 import 'reflect-metadata'
 
-import initApp from './utils/initApp'
-import createWindow from './utils/createWindow'
+import initApp from './startup/initApp'
+import createWindow from './startup/createWindow'
 import i18n from './utils/i18n'
 import mainmenu from './utils/mainmenu'
 import createLoopTask from './startup/loopTask/create'
